requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.9.30=h06a4308_1
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.6=pypi_0
cloudpickle=1.6.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
cython=0.29.24=pypi_0
debugpy=1.4.3=pypi_0
fasteners=0.16.3=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
glfw=2.1.0=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
imageio=2.9.0=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lockfile=0.12.2=pypi_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.4.3=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
mujoco-py=2.0.2.13=dev_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.3.56=pypi_0
openh264=2.1.0=hd408876_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
pillow=8.3.1=py38h2c7a002_0
pip=21.0.1=py38h06a4308_0
pycparser=2.20=pypi_0
pyglet=1.5.20=pypi_0
pyparsing=2.4.7=pypi_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pypi_0
pytorch=1.9.1=py3.8_cuda10.2_cudnn7.6.5_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
scikit-learn=1.0=pypi_0
scipy=1.7.1=pypi_0
setuptools=58.0.3=pypi_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.1=py38
torchvision=0.10.1=py38_cu102
tqdm=4.62.2=pypi_0
typing_extensions=3.10.0.0=pyhca03da5_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
